fix: Move type & rule to nodejs module
